Revert/fix u64/usize type mismatches

+++ src/internal_events/codecs.rs
@@ -104,7 +104,7 @@ impl<'a> InternalEvent for EncoderSerializeError<'a> {
 #[derive(Debug)]
 pub struct EncoderWriteError<'a, E> {
     pub error: &'a E,
-    pub count: usize,
+    pub count: u64,
 }
 
 impl<E: std::fmt::Display> InternalEvent for EncoderWriteError<'_, E> {
